refactor: move delegator vote proof to `penumbra-governance`
