+++ crates/core/component/governance/Cargo.toml
@@ -29,12 +29,20 @@ penumbra-sct = { path = "../sct", default-features = false }
 penumbra-component = { path = "../component", optional = true }
 penumbra-chain = { path = "../chain", default-features = false }
 penumbra-asset = { path = "../../../core/asset", default-features = false  } 
+penumbra-keys = { path = "../../../core/keys", default-features = false  } 
 
 # Penumbra dependencies
 decaf377-rdsa = { version = "0.6" }
+decaf377 = {version = "0.4", features = ["r1cs"] }
 
 # Crates.io dependencies
+base64 = "0.20"
+ark-r1cs-std = {version = "0.4", default-features = false }
+ark-relations = "0.4"
 ark-ff = { version = "0.4", default_features = false }
+ark-serialize = "0.4"
+ark-groth16 = {version = "0.4", default-features = false}
+ark-snark = "0.4"
 metrics = "0.19.0"
 serde = { version = "1", features = ["derive"] }
 tracing = "0.1"
@@ -47,3 +55,6 @@ rand_core = { version = "0.6.3", features = ["getrandom"] }
 rand = "0.8"
 im = "15.1"
 regex = "1.5"
+
+[dev-dependencies]
+proptest = "1"
