+++ crates/core/component/governance/src/delegator_vote.rs
@@ -1,0 +1,1 @@
+pub mod proof;
