+++ crates/core/crypto/src/proofs/groth16.rs
@@ -1,11 +1,9 @@
-mod delegator_vote;
 mod gadgets;
 mod nullifier_derivation;
 mod output;
 mod spend;
 mod traits;
 
-pub use delegator_vote::{DelegatorVoteCircuit, DelegatorVoteProof};
 pub use nullifier_derivation::{NullifierDerivationCircuit, NullifierDerivationProof};
 pub use output::{OutputCircuit, OutputProof};
 pub use spend::{SpendCircuit, SpendProof};
@@ -97,76 +95,6 @@ mod tests {
                 let proof_result = proof.verify(&vk, position, note, nullifier);
 
                 assert!(proof_result.is_ok());
-        }
-    }
-
-    proptest! {
-    #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    fn delegator_vote_happy_path(seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), value_amount in 1..2000000000u64, num_commitments in 0..2000u64) {
-        let (pk, vk) = DelegatorVoteCircuit::generate_prepared_test_parameters();
-        let mut rng = OsRng;
-
-        let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-        let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-        let fvk_sender = sk_sender.full_viewing_key();
-        let ivk_sender = fvk_sender.incoming();
-        let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-        let value_to_send = Value {
-            amount: value_amount.into(),
-            asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-        };
-
-        let note = Note::generate(&mut rng, &sender, value_to_send);
-        let note_commitment = note.commit();
-        let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-        let nk = *sk_sender.nullifier_key();
-        let ak: VerificationKey<SpendAuth> = sk_sender.spend_auth_key().into();
-        let mut sct = tct::Tree::new();
-
-        // Next, we simulate the case where the SCT is not empty by adding `num_commitments`
-        // unrelated items in the SCT.
-        for _ in 0..num_commitments {
-            let random_note_commitment = Note::generate(&mut rng, &sender, value_to_send).commit();
-            sct.insert(tct::Witness::Keep, random_note_commitment).unwrap();
-        }
-
-        sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-        let anchor = sct.root();
-        let state_commitment_proof = sct.witness(note_commitment).unwrap();
-        sct.end_epoch().unwrap();
-
-        let first_note_commitment = Note::generate(&mut rng, &sender, value_to_send).commit();
-        sct.insert(tct::Witness::Keep, first_note_commitment).unwrap();
-        let start_position = sct.witness(first_note_commitment).unwrap().position();
-
-        let balance_commitment = value_to_send.commit(Fr::from(0u64));
-        let rk: VerificationKey<SpendAuth> = rsk.into();
-        let nf = Nullifier::derive(&nk, state_commitment_proof.position(), &note_commitment);
-
-        let blinding_r = Fq::rand(&mut OsRng);
-        let blinding_s = Fq::rand(&mut OsRng);
-
-        let proof = DelegatorVoteProof::prove(
-            blinding_r,
-            blinding_s,
-            &pk,
-            state_commitment_proof,
-            note,
-            spend_auth_randomizer,
-            ak,
-            nk,
-            anchor,
-            balance_commitment,
-            nf,
-            rk,
-            start_position,
-        )
-        .expect("can create proof");
-
-        let proof_result = proof.verify(&vk, anchor, balance_commitment, nf, rk, start_position);
-        assert!(proof_result.is_ok());
         }
     }
