+++ crates/core/transaction/src/action/delegator_vote.rs
@@ -3,8 +3,8 @@ use ark_ff::Zero;
 use decaf377::Fr;
 use decaf377_rdsa::{Signature, SpendAuth, VerificationKey};
 use penumbra_asset::{balance, Value};
-use penumbra_crypto::{proofs::groth16::DelegatorVoteProof, Nullifier};
-use penumbra_governance::VotingReceiptToken;
+use penumbra_crypto::Nullifier;
+use penumbra_governance::{DelegatorVoteProof, VotingReceiptToken};
 use penumbra_num::Amount;
 use penumbra_proto::{core::governance::v1alpha1 as pb, DomainType, TypeUrl};
 use penumbra_tct as tct;
