+++ crates/core/transaction/src/plan/action/delegator_vote.rs
@@ -2,8 +2,8 @@ use ark_ff::UniformRand;
 use decaf377::{FieldExt, Fq, Fr};
 use decaf377_rdsa::{Signature, SpendAuth};
 use penumbra_asset::{Balance, Value};
-use penumbra_crypto::{proofs::groth16::DelegatorVoteProof, Note, Nullifier};
-use penumbra_governance::VotingReceiptToken;
+use penumbra_crypto::{Note, Nullifier};
+use penumbra_governance::{DelegatorVoteProof, VotingReceiptToken};
 use penumbra_keys::FullViewingKey;
 use penumbra_num::Amount;
 use penumbra_proof_params::DELEGATOR_VOTE_PROOF_PROVING_KEY;
