+++ crates/crypto/proof-params/Cargo.toml
@@ -36,6 +36,7 @@ penumbra-tct = { path = "../tct/", features = ["r1cs"] }
 criterion = { version = "0.4", features = ["html_reports"] }
 penumbra-dex = { path = "../../core/component/dex/" }
 penumbra-stake = { path = "../../core/component/stake/" }
+penumbra-governance = { path = "../../core/component/governance/" }
 penumbra-fee = { path = "../../core/component/fee/" }
 
 [build-dependencies]
